Moved ContributeView and AspectsView to src/views
